job scheduler env and pod affinity
