+++ pkg/apis/deployment/jobschedulercomponent.go
@@ -46,7 +46,9 @@ func (js *jobSchedulerComponent) GetPorts() []v1.ComponentPort {
 
 func (js *jobSchedulerComponent) GetEnvironmentVariables() *v1.EnvVarsMap {
 	return &v1.EnvVarsMap{
-		defaults.RadixDeploymentEnvironmentVariable: js.radixDeployment.Name,
+		defaults.RadixDeploymentEnvironmentVariable:               js.radixDeployment.Name,
+		defaults.OperatorEnvLimitDefaultCPUEnvironmentVariable:    os.Getenv(defaults.OperatorEnvLimitDefaultCPUEnvironmentVariable),
+		defaults.OperatorEnvLimitDefaultMemoryEnvironmentVariable: os.Getenv(defaults.OperatorEnvLimitDefaultMemoryEnvironmentVariable),
 	}
 }
